Oh, CMake.
This is getting to the code/comments ratio levels usual for EGL/WGL. Not
fun at all, especially given that I had to bisect by downloading 9 CMake
releases to find the first one that fails, and then only because I
spotted CMP0077 in the 3.13 changelog I got an idea what might have
caused this.

Time spent on this problem so far: 6 hours 30 minutes.