Emscripten: use -flto instead of --llvm-lto.
The old one was specific to the fastcomp backend, which is the default
since 1.39 and is no longer there as of Emscripten 2.0.

Which in turn means the toolchain doesn't support the fastcomp backend
anymore, but that should be fine as 1.39 is from October 2019. (Which
also explains why I'm only merging this now, to not have to deal with
any fastcomp compatibility anymore.)

Co-authored-by: Vladimír Vondruš <mosra@centrum.cz>