UseAndroid: don't use the build-tools subdirectory when finding the SDK.
On Arch at least, installing just android-sdk without any
android-sdk-build-tools-* package doesn't create that directory, which
then causes an error that looks like the whole SDK would not be found.

Which is not true. Instead look for the tools/source.properties file
which should hopefully be there, and split a failure of not finding any
build-tools into a separate error message.